requirements.txt
-i https://pypi.org/simple
asgiref==3.2.3
bcrypt==3.1.7
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.0
django-extensions==2.2.8
django-storages[google]==1.9.1
django==3.0.3
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
h11==0.9.0
httptools==0.1.1 ; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
idna==2.9
protobuf==3.11.3
psycopg2-binary==2.8.4
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.19
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0 ; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
uwsgi==2.0.18
websockets==8.1
